requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
python-multipart>=0.0.6
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
anthropic>=0.7.0

